# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ANALYTE",
    "description": "Bazimi determinist dhe verifikimi i shpjegimeve të analizave laboratorike.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get",
        "operationId": "get_documents__document_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Remove",
        "operationId": "remove_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/documents/{document_id}/explanation": {
      "get": {
        "tags": [
          "explanations"
        ],
        "summary": "Explanation",
        "operationId": "explanation_documents__document_id__explanation_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplanationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/verification": {
      "get": {
        "tags": [
          "explanations"
        ],
        "summary": "Verification",
        "operationId": "verification_documents__document_id__verification_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/data": {
      "delete": {
        "tags": [
          "privacy"
        ],
        "summary": "Erase Account",
        "description": "Fshin llogarinë dhe çdo dokument të saj, pas verifikimit të fjalëkalimit.\n\nPas kësaj tokenët e pronarit nuk vlejnë më (seancat shkojnë me llogarinë), prandaj një përsëritje e së njëjtës kërkesë\nkthen 401, jo një gabim tjetër. Fjalëkalimi i gabuar kthen 403 dhe jo 401, që klienti të mos e marrë për seancë të\nskaduar dhe të nisë një rifreskim.",
        "operationId": "erase_account_me_data_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ErasureIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/me/export": {
      "get": {
        "tags": [
          "privacy"
        ],
        "summary": "Export",
        "operationId": "export_me_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttemptOut": {
        "properties": {
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "generator": {
            "type": "string",
            "title": "Generator"
          },
          "is_fallback": {
            "type": "boolean",
            "title": "Is Fallback"
          },
          "delivered": {
            "type": "boolean",
            "title": "Delivered"
          },
          "failed": {
            "type": "boolean",
            "title": "Failed"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "attempt",
          "generator",
          "is_fallback",
          "delivered",
          "failed",
          "verification"
        ],
        "title": "AttemptOut"
      },
...
      "DocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "state_reason": {
            "type": "string",
            "title": "State Reason"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "terminal": {
            "type": "boolean",
            "title": "Terminal"
          },
          "model_consent": {
            "type": "boolean",
            "title": "Model Consent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "uploaded_at",
          "size_bytes",
          "state",
          "state_reason",
          "channel",
          "terminal",
          "model_consent"
        ],
        "title": "DocumentOut"
      },
...
      "ErasureIn": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "ErasureIn"
      },
      "ExplanationOut": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "generator": {
            "type": "string",
            "title": "Generator"
          },
          "critical": {
            "type": "boolean",
            "title": "Critical"
          },
          "banner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banner"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          },
          "notices": {
            "items": {
              "$ref": "#/components/schemas/Notice"
            },
            "type": "array",
            "title": "Notices"
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationSummary"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "text",
          "generator",
          "critical",
          "banner",
          "disclaimer",
          "notices",
          "verification"
        ],
        "title": "ExplanationOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Notice": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "code",
          "text"
        ],
        "title": "Notice",
        "description": "Një kufizim i shprehur nga sistemi. `code` e lejon ndërfaqen ta vendosë\naty ku duhet — njoftimi i OCR-së në krye, jo i humbur në fund."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationDetail": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "rules_version": {
            "type": "string",
            "title": "Rules Version"
          },
          "classifier_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classifier Version"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/ViolationOut"
            },
            "type": "array",
            "title": "Violations"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "mode",
          "rules_version",
          "classifier_version",
          "duration_ms",
          "violations"
        ],
        "title": "VerificationDetail"
      },
      "VerificationOut": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationSummary"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/AttemptOut"
            },
            "type": "array",
            "title": "Attempts"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "verification",
          "attempts"
        ],
        "title": "VerificationOut"
      },
      "VerificationSummary": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "violation_count": {
            "type": "integer",
            "title": "Violation Count"
          },
          "is_fallback": {
            "type": "boolean",
            "title": "Is Fallback"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "mode",
          "violation_count",
          "is_fallback"
        ],
        "title": "VerificationSummary"
      },
      "ViolationOut": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "detected_by": {
            "type": "string",
            "title": "Detected By"
          },
          "sentence": {
            "type": "string",
            "title": "Sentence"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "type",
          "detected_by",
          "sentence",
          "evidence",
          "confidence"
        ],
        "title": "ViolationOut"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}